+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/CameraAndroidCameraxPlugin.java
@@ -27,6 +27,7 @@ public final class CameraAndroidCameraxPlugin implements FlutterPlugin, Activity
   @VisibleForTesting @Nullable public ImageCaptureHostApiImpl imageCaptureHostApiImpl;
   @VisibleForTesting @Nullable public CameraControlHostApiImpl cameraControlHostApiImpl;
   @VisibleForTesting @Nullable public SystemServicesHostApiImpl systemServicesHostApiImpl;
+  @VisibleForTesting @Nullable public MeteringPointHostApiImpl meteringPointHostApiImpl;
 
   @VisibleForTesting @Nullable
   public Camera2CameraControlHostApiImpl camera2CameraControlHostApiImpl;
@@ -127,6 +128,12 @@ public final class CameraAndroidCameraxPlugin implements FlutterPlugin, Activity
         binaryMessenger, camera2CameraControlHostApiImpl);
     GeneratedCameraXLibrary.CaptureRequestOptionsHostApi.setup(
         binaryMessenger, new CaptureRequestOptionsHostApiImpl(instanceManager));
+    GeneratedCameraXLibrary.FocusMeteringActionHostApi.setup(
+        binaryMessenger, new FocusMeteringActionHostApiImpl(instanceManager));
+    GeneratedCameraXLibrary.FocusMeteringResultHostApi.setup(
+        binaryMessenger, new FocusMeteringResultHostApiImpl(instanceManager));
+    meteringPointHostApiImpl = new MeteringPointHostApiImpl(instanceManager);
+    GeneratedCameraXLibrary.MeteringPointHostApi.setup(binaryMessenger, meteringPointHostApiImpl);
   }
 
   @Override
@@ -248,6 +255,9 @@ public final class CameraAndroidCameraxPlugin implements FlutterPlugin, Activity
     }
     if (deviceOrientationManagerHostApiImpl != null) {
       deviceOrientationManagerHostApiImpl.setActivity(activity);
+    }
+    if (meteringPointHostApiImpl != null) {
+      meteringPointHostApiImpl.setActivity(activity);
     }
   }
 }