+++ packages/camera/camera_android_camerax/android/src/test/java/io/flutter/plugins/camerax/CameraAndroidCameraxPluginTest.java
@@ -92,6 +92,8 @@ public class CameraAndroidCameraxPluginTest {
         mock(SystemServicesHostApiImpl.class);
     final DeviceOrientationManagerHostApiImpl mockDeviceOrientationManagerHostApiImpl =
         mock(DeviceOrientationManagerHostApiImpl.class);
+    final MeteringPointHostApiImpl mockMeteringPointHostApiImpl =
+        mock(MeteringPointHostApiImpl.class);
     final ArgumentCaptor<PermissionsRegistry> permissionsRegistryCaptor =
         ArgumentCaptor.forClass(PermissionsRegistry.class);
 
@@ -103,6 +105,7 @@ public class CameraAndroidCameraxPluginTest {
     plugin.liveDataHostApiImpl = mock(LiveDataHostApiImpl.class);
     plugin.systemServicesHostApiImpl = mockSystemServicesHostApiImpl;
     plugin.deviceOrientationManagerHostApiImpl = mockDeviceOrientationManagerHostApiImpl;
+    plugin.meteringPointHostApiImpl = mockMeteringPointHostApiImpl;
 
     plugin.onAttachedToEngine(flutterPluginBinding);
     plugin.onAttachedToActivity(activityPluginBinding);
@@ -110,6 +113,7 @@ public class CameraAndroidCameraxPluginTest {
     // Check Activity references are set.
     verify(mockSystemServicesHostApiImpl).setActivity(mockActivity);
     verify(mockDeviceOrientationManagerHostApiImpl).setActivity(mockActivity);
+    verify(mockMeteringPointHostApiImpl).setActivity(mockActivity);
 
     // Check permissions registry reference is set.
     verify(mockSystemServicesHostApiImpl)
@@ -129,11 +133,14 @@ public class CameraAndroidCameraxPluginTest {
         mock(SystemServicesHostApiImpl.class);
     final DeviceOrientationManagerHostApiImpl mockDeviceOrientationManagerHostApiImpl =
         mock(DeviceOrientationManagerHostApiImpl.class);
+    final MeteringPointHostApiImpl mockMeteringPointHostApiImpl =
+        mock(MeteringPointHostApiImpl.class);
 
     plugin.processCameraProviderHostApiImpl = mockProcessCameraProviderHostApiImpl;
     plugin.liveDataHostApiImpl = mockLiveDataHostApiImpl;
     plugin.systemServicesHostApiImpl = mockSystemServicesHostApiImpl;
     plugin.deviceOrientationManagerHostApiImpl = mockDeviceOrientationManagerHostApiImpl;
+    plugin.meteringPointHostApiImpl = mockMeteringPointHostApiImpl;
 
     plugin.onAttachedToEngine(flutterPluginBinding);
     plugin.onDetachedFromActivityForConfigChanges();
@@ -142,6 +149,7 @@ public class CameraAndroidCameraxPluginTest {
     verify(mockLiveDataHostApiImpl).setLifecycleOwner(null);
     verify(mockSystemServicesHostApiImpl).setActivity(null);
     verify(mockDeviceOrientationManagerHostApiImpl).setActivity(null);
+    verify(mockMeteringPointHostApiImpl).setActivity(null);
   }
 
   @Test
@@ -257,6 +265,8 @@ public class CameraAndroidCameraxPluginTest {
         mock(DeviceOrientationManagerHostApiImpl.class);
     final Camera2CameraControlHostApiImpl mockCamera2CameraControlHostApiImpl =
         mock(Camera2CameraControlHostApiImpl.class);
+    final MeteringPointHostApiImpl mockMeteringPointHostApiImpl =
+        mock(MeteringPointHostApiImpl.class);
     final ArgumentCaptor<PermissionsRegistry> permissionsRegistryCaptor =
         ArgumentCaptor.forClass(PermissionsRegistry.class);
 
@@ -271,6 +281,7 @@ public class CameraAndroidCameraxPluginTest {
     plugin.imageAnalysisHostApiImpl = mockImageAnalysisHostApiImpl;
     plugin.cameraControlHostApiImpl = mockCameraControlHostApiImpl;
     plugin.deviceOrientationManagerHostApiImpl = mockDeviceOrientationManagerHostApiImpl;
+    plugin.meteringPointHostApiImpl = mockMeteringPointHostApiImpl;
     plugin.liveDataHostApiImpl = mock(LiveDataHostApiImpl.class);
     plugin.camera2CameraControlHostApiImpl = mockCamera2CameraControlHostApiImpl;
 
@@ -280,6 +291,7 @@ public class CameraAndroidCameraxPluginTest {
     // Check Activity references are set.
     verify(mockSystemServicesHostApiImpl).setActivity(mockActivity);
     verify(mockDeviceOrientationManagerHostApiImpl).setActivity(mockActivity);
+    verify(mockMeteringPointHostApiImpl).setActivity(mockActivity);
 
     // Check Activity as Context references are set.
     verify(mockProcessCameraProviderHostApiImpl).setContext(mockActivity);
@@ -308,11 +320,14 @@ public class CameraAndroidCameraxPluginTest {
     final LiveDataHostApiImpl mockLiveDataHostApiImpl = mock(LiveDataHostApiImpl.class);
     final DeviceOrientationManagerHostApiImpl mockDeviceOrientationManagerHostApiImpl =
         mock(DeviceOrientationManagerHostApiImpl.class);
+    final MeteringPointHostApiImpl mockMeteringPointHostApiImpl =
+        mock(MeteringPointHostApiImpl.class);
 
     plugin.processCameraProviderHostApiImpl = mockProcessCameraProviderHostApiImpl;
     plugin.liveDataHostApiImpl = mockLiveDataHostApiImpl;
     plugin.systemServicesHostApiImpl = mockSystemServicesHostApiImpl;
     plugin.deviceOrientationManagerHostApiImpl = mockDeviceOrientationManagerHostApiImpl;
+    plugin.meteringPointHostApiImpl = mockMeteringPointHostApiImpl;
 
     plugin.onAttachedToEngine(flutterPluginBinding);
     plugin.onDetachedFromActivityForConfigChanges();
@@ -321,6 +336,7 @@ public class CameraAndroidCameraxPluginTest {
     verify(mockLiveDataHostApiImpl).setLifecycleOwner(null);
     verify(mockSystemServicesHostApiImpl).setActivity(null);
     verify(mockDeviceOrientationManagerHostApiImpl).setActivity(null);
+    verify(mockMeteringPointHostApiImpl).setActivity(null);
   }
 
   @Test