+++ packages/camera/camera_android_camerax/lib/src/camerax_proxy.dart
@@ -7,15 +7,18 @@ import 'dart:ui' show Size;
 import 'analyzer.dart';
 import 'camera2_camera_control.dart';
 import 'camera_control.dart';
+import 'camera_info.dart';
 import 'camera_selector.dart';
 import 'camera_state.dart';
 import 'camerax_library.g.dart';
 import 'capture_request_options.dart';
 import 'device_orientation_manager.dart';
 import 'fallback_strategy.dart';
+import 'focus_metering_action.dart';
 import 'image_analysis.dart';
 import 'image_capture.dart';
 import 'image_proxy.dart';
+import 'metering_point.dart';
 import 'observer.dart';
 import 'preview.dart';
 import 'process_camera_provider.dart';
@@ -54,6 +57,8 @@ class CameraXProxy {
     this.getDefaultDisplayRotation = _getDefaultDisplayRotation,
     this.getCamera2CameraControl = _getCamera2CameraControl,
     this.createCaptureRequestOptions = _createCaptureRequestOptions,
+    this.createMeteringPoint = _createMeteringPoint,
+    this.createFocusMeteringAction = _createFocusMeteringAction,
   });
 
   /// Returns a [ProcessCameraProvider] instance.
@@ -150,6 +155,16 @@ class CameraXProxy {
   CaptureRequestOptions Function(
           List<(CaptureRequestKeySupportedType, Object?)> options)
       createCaptureRequestOptions;
+
+  /// Returns a [MeteringPoint] with the specified coordinates based on
+  /// [cameraInfo].
+  MeteringPoint Function(double x, double y, CameraInfo cameraInfo)
+      createMeteringPoint;
+
+  /// Returns a [FocusMeteringAction] based on the specified metering points
+  /// and their modes.
+  FocusMeteringAction Function(List<(MeteringPoint, int?)> meteringPointInfos)
+      createFocusMeteringAction;
 
   static Future<ProcessCameraProvider> _getProcessCameraProvider() {
     return ProcessCameraProvider.getInstance();
@@ -262,5 +277,15 @@ class CameraXProxy {
   static CaptureRequestOptions _createCaptureRequestOptions(
       List<(CaptureRequestKeySupportedType, Object?)> options) {
     return CaptureRequestOptions(requestedOptions: options);
+  }
+
+  static MeteringPoint _createMeteringPoint(
+      double x, double y, CameraInfo cameraInfo) {
+    return MeteringPoint(x: x, y: y, cameraInfo: cameraInfo);
+  }
+
+  static FocusMeteringAction _createFocusMeteringAction(
+      List<(MeteringPoint, int?)> meteringPointInfos) {
+    return FocusMeteringAction(meteringPointInfos: meteringPointInfos);
   }
 }