+++ packages/in_app_purchase/in_app_purchase_storekit/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.11
+
+* Fixes SKError.userInfo not being nullable.
+
+## 0.3.10
+
+* Converts `startProductRequest()`, `finishTransaction()`, `restoreTransactions()`, `presentCodeRedemptionSheet()` to pigeon.
+
 ## 0.3.9
 
 * Converts `storefront()`, `transactions()`, `addPayment()`, `canMakePayment` to pigeon.